refactor sql storage and indexeddb tests
