+++ mm2src/coins/utxo/utxo_block_header_storage/wasm/indexeddb_block_header_storage.rs
@@ -279,163 +279,14 @@ impl BlockHeaderStorageOps for IDBBlockHeadersStorage {
     }
 }
 
-#[cfg(target_arch = "wasm32")]
-mod tests {
-    use super::IDBBlockHeadersStorage;
-    use super::*;
-    use chain::BlockHeaderBits;
-    use common::log::info;
-    use common::log::wasm_log::register_wasm_log;
-    use mm2_test_helpers::for_tests::mm_ctx_with_custom_db;
-    use wasm_bindgen_test::{wasm_bindgen_test, wasm_bindgen_test_configure};
+#[cfg(test)]
+impl IDBBlockHeadersStorage {
+    pub(crate) async fn is_table_empty(&self, _table_name: &str) -> bool {
+        let ticker = self.ticker.clone();
+        let locked_db = self.lock_db().await.unwrap();
+        let db_transaction = locked_db.get_inner().transaction().await.unwrap();
+        let block_headers_db = db_transaction.table::<BlockHeaderStorageTable>().await.unwrap();
 
-    wasm_bindgen_test_configure!(run_in_browser);
-
-    #[wasm_bindgen_test]
-    async fn test_storage_init() {
-        let ctx = mm_ctx_with_custom_db();
-        let storage = IDBBlockHeadersStorage::new(&ctx, "RICK".to_string());
-
-        register_wasm_log();
-
-        let initialized = storage.is_initialized_for().await.unwrap();
-        assert!(initialized);
-
-        // repetitive init must not fail
-        storage.init().await.unwrap();
-
-        let initialized = storage.is_initialized_for().await.unwrap();
-        assert!(initialized);
-    }
-
-    #[wasm_bindgen_test]
-    async fn test_add_block_headers() {
-        let ctx = mm_ctx_with_custom_db();
-        let storage = IDBBlockHeadersStorage::new(&ctx, "RICK".to_string());
-        register_wasm_log();
-
-        let mut headers = HashMap::with_capacity(1);
-        let block_header: BlockHeader = "0000002076d41d3e4b0bfd4c0d3b30aa69fdff3ed35d85829efd04000000000000000000b386498b583390959d9bac72346986e3015e83ac0b54bc7747a11a494ac35c94bb3ce65a53fb45177f7e311c".into();
-        headers.insert(520481, block_header);
-        let add_headers = storage.add_block_headers_to_storage(headers).await;
-
-        assert!(add_headers.is_ok());
-    }
-
-    #[wasm_bindgen_test]
-    async fn test_get_block_header() {
-        let ctx = mm_ctx_with_custom_db();
-        let storage = IDBBlockHeadersStorage::new(&ctx, "RICK".to_string());
-        register_wasm_log();
-
-        let mut headers = HashMap::with_capacity(1);
-        let block_header: BlockHeader = "0000002076d41d3e4b0bfd4c0d3b30aa69fdff3ed35d85829efd04000000000000000000b386498b583390959d9bac72346986e3015e83ac0b54bc7747a11a494ac35c94bb3ce65a53fb45177f7e311c".into();
-        headers.insert(520481, block_header);
-        let add_headers = storage.add_block_headers_to_storage(headers).await;
-        assert!(add_headers.is_ok());
-
-        let block_header_hex = storage.get_block_header_raw(520481).await.unwrap();
-        assert_eq!
-        ("0000002076d41d3e4b0bfd4c0d3b30aa69fdff3ed35d85829efd04000000000000000000b386498b583390959d9bac72346986e3015e83ac0b54bc7747a11a494ac35c94bb3ce65a53fb45177f7e311c".to_string(), block_header_hex.unwrap());
-
-        let expected_header_hash: H256 = "0000000000000000002e31d0714a5ab23100945ff87ba2d856cd566a3c9344ec".into();
-        let block_header = storage.get_block_header(520481).await.unwrap().unwrap();
-        assert_eq!(block_header.hash(), expected_header_hash.reversed());
-    }
-
-    #[wasm_bindgen_test]
-    async fn test_get_block_header_raw() {
-        let ctx = mm_ctx_with_custom_db();
-        let storage = IDBBlockHeadersStorage::new(&ctx, "RICK".to_string());
-        register_wasm_log();
-
-        let mut headers = HashMap::with_capacity(1);
-        let block_header: BlockHeader = "02000000ea01a61a2d7420a1b23875e40eb5eb4ca18b378902c8e6384514ad0000000000c0c5a1ae80582b3fe319d8543307fa67befc2a734b8eddb84b1780dfdf11fa2b20e71353ffff001d00805fe0".into();
-        headers.insert(201595, block_header.clone());
-        let add_headers = storage.add_block_headers_to_storage(headers).await;
-        assert!(add_headers.is_ok());
-
-        let raw_header = storage.get_block_header_raw(201595).await.unwrap();
-        assert_eq!("02000000ea01a61a2d7420a1b23875e40eb5eb4ca18b378902c8e6384514ad0000000000c0c5a1ae80582b3fe319d8543307fa67befc2a734b8eddb84b1780dfdf11fa2b20e71353ffff001d00805fe0", raw_header.unwrap())
-    }
-
-    #[wasm_bindgen_test]
-    async fn test_get_last_block_height() {
-        let ctx = mm_ctx_with_custom_db();
-        let storage = IDBBlockHeadersStorage::new(&ctx, "RICK".to_string());
-        register_wasm_log();
-
-        let mut headers = HashMap::with_capacity(1);
-        // This block has max difficulty
-        // https://live.blockcypher.com/btc-testnet/block/00000000961a9d117feb57e516e17217207a849bf6cdfce529f31d9a96053530/
-        let block_header: BlockHeader = "02000000ea01a61a2d7420a1b23875e40eb5eb4ca18b378902c8e6384514ad0000000000c0c5a1ae80582b3fe319d8543307fa67befc2a734b8eddb84b1780dfdf11fa2b20e71353ffff001d00805fe0".into();
-        headers.insert(201595, block_header);
-
-        // https://live.blockcypher.com/btc-testnet/block/0000000000ad144538e6c80289378ba14cebb50ee47538b2a120742d1aa601ea/
-        let expected_block_header: BlockHeader = "02000000cbed7fd98f1f06e85c47e13ff956533642056be45e7e6b532d4d768f00000000f2680982f333fcc9afa7f9a5e2a84dc54b7fe10605cd187362980b3aa882e9683be21353ab80011c813e1fc0".into();
-        headers.insert(201594, expected_block_header.clone());
-
-        // This block has max difficulty
-        // https://live.blockcypher.com/btc-testnet/block/0000000000ad144538e6c80289378ba14cebb50ee47538b2a120742d1aa601ea/
-        let block_header: BlockHeader = "020000001f38c8e30b30af912fbd4c3e781506713cfb43e73dff6250348e060000000000afa8f3eede276ccb4c4ee649ad9823fc181632f262848ca330733e7e7e541beb9be51353ffff001d00a63037".into();
-
-        headers.insert(201593, block_header);
-        let add_headers = storage.add_block_headers_to_storage(headers.clone()).await;
-        assert!(add_headers.is_ok());
-
-        let last_height = storage.get_last_block_height().await.unwrap();
-        assert_eq!(201595, last_height);
-    }
-
-    #[wasm_bindgen_test]
-    async fn test_get_last_block_header_with_non_max_bits() {
-        let ctx = mm_ctx_with_custom_db();
-        let storage = IDBBlockHeadersStorage::new(&ctx, "RICK".to_string());
-        register_wasm_log();
-        let mut headers = HashMap::with_capacity(2);
-
-        // This block has max difficulty
-        // https://live.blockcypher.com/btc-testnet/block/00000000961a9d117feb57e516e17217207a849bf6cdfce529f31d9a96053530/
-        let block_header: BlockHeader = "02000000ea01a61a2d7420a1b23875e40eb5eb4ca18b378902c8e6384514ad0000000000c0c5a1ae80582b3fe319d8543307fa67befc2a734b8eddb84b1780dfdf11fa2b20e71353ffff001d00805fe0".into();
-        headers.insert(201595, block_header);
-
-        // https://live.blockcypher.com/btc-testnet/block/0000000000ad144538e6c80289378ba14cebb50ee47538b2a120742d1aa601ea/
-        let expected_block_header: BlockHeader = "02000000cbed7fd98f1f06e85c47e13ff956533642056be45e7e6b532d4d768f00000000f2680982f333fcc9afa7f9a5e2a84dc54b7fe10605cd187362980b3aa882e9683be21353ab80011c813e1fc0".into();
-        headers.insert(201594, expected_block_header.clone());
-
-        // This block has max difficulty
-        // https://live.blockcypher.com/btc-testnet/block/0000000000ad144538e6c80289378ba14cebb50ee47538b2a120742d1aa601ea/
-        let block_header: BlockHeader = "020000001f38c8e30b30af912fbd4c3e781506713cfb43e73dff6250348e060000000000afa8f3eede276ccb4c4ee649ad9823fc181632f262848ca330733e7e7e541beb9be51353ffff001d00a63037".into();
-        headers.insert(201593, block_header);
-
-        let add_headers = storage.add_block_headers_to_storage(headers).await;
-        assert!(add_headers.is_ok());
-
-        let actual_block_header = storage
-            .get_last_block_header_with_non_max_bits()
-            .await
-            .unwrap()
-            .unwrap();
-        assert_ne!(actual_block_header.bits, BlockHeaderBits::Compact(MAX_BITS_BTC.into()));
-        assert_eq!(actual_block_header, expected_block_header);
-    }
-
-    #[wasm_bindgen_test]
-    async fn test_get_block_height_by_hash() {
-        let ctx = mm_ctx_with_custom_db();
-        let storage = IDBBlockHeadersStorage::new(&ctx, "RICK".to_string());
-        register_wasm_log();
-
-        let mut headers = HashMap::with_capacity(1);
-        let block_header: BlockHeader = "02000000ea01a61a2d7420a1b23875e40eb5eb4ca18b378902c8e6384514ad0000000000c0c5a1ae80582b3fe319d8543307fa67befc2a734b8eddb84b1780dfdf11fa2b20e71353ffff001d00805fe0".into();
-        headers.insert(201595, block_header.clone());
-        let add_headers = storage.add_block_headers_to_storage(headers).await;
-        assert!(add_headers.is_ok());
-
-        let raw_header = storage
-            .get_block_height_by_hash("00000000961a9d117feb57e516e17217207a849bf6cdfce529f31d9a96053530".into())
-            .await
-            .unwrap();
-        assert_eq!(201595, raw_header.unwrap())
+        block_headers_db.get_items("ticker", ticker).await.unwrap().is_empty()
     }
 }
